requirements.txt
mcp>=1.0.0
httpx>=0.27.1
python-dotenv>=1.0.0

